Support mixing command & folders
To provide support for backups that consist of both a folder & command
output (for example database + files), the file can be saved in the
folder to be backed up. It's required to provide a file name in that
case. The namespace of the command can be different to the volume (for
example with a central database server) so it's possible to specify a
different namespace for that. The original way still works if only
running a command: Using the existing `namespace` option is equivalent.