Update dataLocality
By switching to best-effort I hope it can be a bit faster since it's
more likely the node that runs the backup has the volume on it. I'm not
sure if it'll work due to the dynamic nature and there only being one
replica. It might be better to schedule the pod where the volume is.