Fix rename bug
When trying to rename a file across file system boundaries, Golang
doesn't behave like "mv" and switch to a copy mode, it instead returns
an error. However, it's not necessary to introduce a copy handling.

Instead, we already know the correct destination for files, so why move
it around? Write it directly to its intended destination and only use
the cache file for the case where std input is used for backup.