Add timeout for getting snapshot content
It's possible for this step to hang, such as when having misaligned
storage classes (fixed in previous commit). In that case, the job would
hang forever without this timeout. It's possible that some
implementations will take longer than the current timeout, so it might
need to be re-adjusted. "waitTimeout" is rather short, not meant for
e.g. steps that copy data and might take longer with larger backups.