Add database with last backup run
This introduces a database with a single table: Storing a profile with a
timestamp of the last successful run. When a backup succeeds, the
timestamp is updated. If it was successful within a certain interval,
the backup if skipped. The interval is hardcoded to 12h at the moment.

This feature was introduced because if any single backup run fails, all
of them get re-run which is costly and disruptive, especially for those
services that need to be stopped to be backed up. By remembering which
backups succeeded, it's possible to ensure that only the failed backups
run.

Right now this is a very basic implementation, hopefully good enough to
be useful, with option for future improvement. For example, there's no
support for creating the table (done manually) or migrations. Right now,
the interval isn't even configurable. All of that can be added later.