Add BackupPvc interface
There's now two ways of getting a PVC to take a backup from, both of
which have different internals but very similar behaviour externally.
I've created an interface that abstracts the common expectation away and
preserves their internals.

I moved them into a new sub-package which created a circular dependency
on their parent for the NewObjectMeta call. I don't know if it would
have been an issue, but to keep it cleaner, I moved that function into a
new package `meta` and updated all references.