Fix validation bug for mixed pod finders
With the addition of a third method (selectors) I messed up the
validation logic for multiple ways to find the target pod being set. The
problem was that the boolean condition did not correctly check if
exactly one method was used. Using 2 out of 3 would not fail this check,
but should've.

In addition, the unit tests were broken that were supposed to catch
this, failing on a different error that made tests succeed by accident.

In the fix, I count the number of methods to find pods and report an
error if it isn't exactly 1. I've also fixed the fail condition for the
tests so they work again and would now detect this problem.