Add support for command-before & command-after
These two options allow running a command before and after the backup,
against the pod that was selected, optionally specifying a container
name, otherwise Kubernetes picks it.

I've made a tradeoff here in passing the *rest.Config object around:
It's not an interface and can't be unit tested, but there doesn't seem
to be a fake package for the remotecommand package so I didn't want to
write my own just to provide an interface if I don't have unit tests for
that part yet anyway. In the future when adding unit tests there should
probably be an interface here. Or maybe some encapsulating object that
handles all those things.