Add test for overlapping folders
There's no bug at the moment, but this seems like an edge case that
could easily be missed: If someone mounts one folder on top of another,
then we should return both of them correctly.

The backup profile still needs to include both folder explicitly for
them to be included.